requirements.txt
Django>=2.0
psycopg2>=2.7
djangorestframework
django-filter
djangorestframework-jwt
django-cors-headers
django-filter